+++ src/test/java/com/hunko/missionmatching/helper/FakeMissionSaver.java
@@ -2,9 +2,9 @@ package com.hunko.missionmatching.helper;
 
 
 import com.hunko.missionmatching.core.domain.Mission;
+import com.hunko.missionmatching.core.domain.MissionSaver;
 import com.hunko.missionmatching.storage.MissionEntity;
 import com.hunko.missionmatching.storage.MissionMapper;
-import com.hunko.missionmatching.core.domain.MissionSaver;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -20,7 +20,7 @@ public class FakeMissionSaver extends MissionSaver {
     public Long save(Mission mission) {
         MissionEntity missionEntity = MissionMapper.toMissionEntity(mission);
         MissionEntity copy = MissionEntity.builder()
-                .id(mission.getId() == null ? missions.size() : mission.getId())
+                .id(mission.getId().toLong() == null ? missions.size() : mission.getId().toLong())
                 .title(mission.getTitle())
                 .status(mission.getStatus())
                 .startDate(missionEntity.getStartDate())
@@ -29,10 +29,10 @@ public class FakeMissionSaver extends MissionSaver {
                 .build();
         Mission copyMission = MissionMapper.toMission(copy);
         missions.add(copyMission);
-        return copyMission.getId();
+        return copyMission.getId().toLong();
     }
 
     public Mission getMission(int id) {
-        return missions.stream().filter(mission -> mission.getId() == id).findFirst().orElse(null);
+        return missions.stream().filter(mission -> mission.getId().toLong() == id).findFirst().orElse(null);
     }
 }
